Clean up comments, and move CONTEXT type.
