+++ lib/pbench/server/api/resources/query_apis/datasets_detail.py
@@ -4,14 +4,13 @@ from logging import Logger
 
 from pbench.server import PbenchServerConfig
 from pbench.server.api.resources import (
-    CONTEXT,
     JSON,
     Schema,
     Parameter,
     ParamType,
     PostprocessError,
 )
-from pbench.server.api.resources.query_apis import ElasticBase
+from pbench.server.api.resources.query_apis import CONTEXT, ElasticBase
 
 
 class DatasetsDetail(ElasticBase):
