+++ lib/pbench/server/api/resources/query_apis/datasets_list.py
@@ -2,8 +2,8 @@ from flask import jsonify
 from logging import Logger
 
 from pbench.server import PbenchServerConfig
-from pbench.server.api.resources import CONTEXT, JSON, Schema, Parameter, ParamType
-from pbench.server.api.resources.query_apis import ElasticBase
+from pbench.server.api.resources import JSON, Schema, Parameter, ParamType
+from pbench.server.api.resources.query_apis import CONTEXT, ElasticBase
 
 
 class DatasetsList(ElasticBase):
