+++ lib/pbench/server/api/resources/query_apis/datasets_publish.py
@@ -7,15 +7,13 @@ from flask_restful import abort
 from pbench.server import PbenchServerConfig
 from pbench.server.api.resources import (
     API_OPERATION,
-    CONTEXT,
     JSON,
     Schema,
     Parameter,
     ParamType,
     PostprocessError,
 )
-from pbench.server.api.resources.query_apis import ElasticBase
-
+from pbench.server.api.resources.query_apis import CONTEXT, ElasticBase
 from pbench.server.database.models.datasets import Dataset, Metadata
 from pbench.server.database.models.users import User
